feat: Added middleware for post owners
Created middleware that will only allow the owner of a post or a post
admin to edit or delete the post.

+++ routes/web.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\FriendController;
 use App\Http\Controllers\PostController;
 use App\Http\Controllers\CommentController;
+use App\Http\Middleware\PostOwner;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
@@ -56,7 +57,7 @@ Route::name("friend.")->prefix("/friend")->controller(FriendController::class)->
     Route::patch("/accept/{id}", "acceptRequest")->name("accept");
     Route::delete("/reject/{id}", "rejectRequest")->name("reject");
     Route::delete("/remove/{id}", "removeFriend")->name("remove");
-});
+})->middleware(["auth", "verified"]);
 
 Route::name("post.")->prefix("/post")->controller(PostController::class)->group(function(){
     Route::get("/index", "index")->name("index");
@@ -64,33 +65,33 @@ Route::name("post.")->prefix("/post")->controller(PostController::class)->group(
     Route::get("/create", "create")->name("create");
     Route::post("", "store")->name("store");
     Route::get("/{id}/show", "show")->name("show");
-    Route::get("/{id}/edit", "edit")->name("edit");
-    Route::put("/{id}", "update")->name("update");
-    Route::delete("/{id}", "destroy")->name("destroy");
+    Route::get("/{id}/edit", "edit")->name("edit")->middleware(["PostOwner"]);
+    Route::put("/{id}", "update")->name("update")->middleware(["PostOwner"]);
+    Route::delete("/{id}", "destroy")->name("destroy")->middleware(["PostOwner"]);
     Route::get("/{id}/likes", "likesByPost")->name("byPost");
-    Route::get("/{id}/saved", "whoSavedPost")->name("whoSaved");
+    Route::get("/{id}/saved", "whoSavedPost")->name("whoSaved")->middleware(["PostOwner"]);
 });
 
 Route::name("like.")->prefix("/like")->controller(LikeController::class)->group(function(){
     Route::put("/{id}", "likePost")->name("add");
     Route::delete("/{id}", "removeLike")->name("remove");
     Route::get("/user", "yourLikes")->name("yours");
-});
+})->middleware(["auth", "verified"]);
 
 Route::name("save.")->prefix("/save")->controller(SaveController::class)->group(function(){
     Route::post("/{id}", "savePost")->name("post");
     Route::delete("/{id}", "unsavePost")->name("remove");
     Route::get("/user", "yourSaves")->name("yours");
-});
+})->middleware(["auth", "verified"]);
 
 Route::name("comment.")->prefix("/comment")->controller(CommentController::class)->group(function(){
     Route::post("/{id}", "store")->name("store");
     Route::delete("/{id}", "destroy")->name("delete");
     Route::get("/{id}/show", "show")->name("show");
-});
+})->middleware(["auth", "verified"]);
 
 Route::name("reply.")->prefix("/reply")->controller(ReplyController::class)->group(function(){
     Route::post("/{id}", "store")->name("store");
     Route::delete("'/{id}", "destroy")->name("delete");
-});
+})->middleware(["auth", "verified"]);
 require __DIR__.'/auth.php';
